+++ secretspec/src/bin/git-credential-secretspec.rs
@@ -1,0 +1,5 @@
+use miette::Result;
+
+fn main() -> Result<()> {
+    secretspec::git_credential::main()
+}